fix: restore previous API
