+++ tests/test_pdfminer_crypto.py
@@ -1,11 +1,10 @@
 """Test of various compression/encoding modules (previously in doctests)"""
 
 import binascii
-from base64 import a85decode
 
 from pdfminer.arcfour import Arcfour
+from pdfminer.ascii85 import ascii85decode, asciihexdecode
 from pdfminer.lzw import lzwdecode
-from pdfminer.pdftypes import asciihexdecode
 from pdfminer.runlength import rldecode
 
 
@@ -19,13 +18,13 @@ def dehex(b):
     return binascii.unhexlify(b)
 
 
-class TestAsciiHex:
+class TestAscii85:
     def test_ascii85decode(self):
         """The sample string is taken from:
         http://en.wikipedia.org/w/index.php?title=Ascii85
         """
-        assert a85decode(b"9jqo^BlbD-BleB1DJ+*+F(f,q") == b"Man is distinguished"
-        assert a85decode(b"E,9)oF*2M7/c~>", adobe=True) == b"pleasure."
+        assert ascii85decode(b"9jqo^BlbD-BleB1DJ+*+F(f,q") == b"Man is distinguished"
+        assert ascii85decode(b"E,9)oF*2M7/c~>") == b"pleasure."
 
     def test_asciihexdecode(self):
         assert asciihexdecode(b"61 62 2e6364   65") == b"ab.cde"
